# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для работы с фильмами и актерами в FilmHub. login: admin, password: kek",
        "title": "FilmHub API",
        "contact": {
            "url": "http://www.github.com/avealice"
        }
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
        "/api/actor": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает нового актера.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/actor"
                ],
                "summary": "Создать актера.",
                "parameters": [
                    {
                        "description": "Данные нового актера",
                        "name": "actor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputActor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Актер успешно создан",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/actor/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получает информацию об актере по его идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/actor/{id}"
                ],
                "summary": "Получить информацию об актере.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор актера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ActorWithMovies"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию об актере по его идентификатору.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/actor/{id}"
                ],
                "summary": "Обновить информацию об актере.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор актера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные актера",
                        "name": "actor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputActor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об актере успешно обновлена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет актера по его идентификатору.",
                "tags": [
                    "/api/actor/{id}"
                ],
                "summary": "Удалить актера.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор актера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Актер успешно удален",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/actors": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить всех актеров из базы данных.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/actors"
                ],
                "summary": "Получить всех актеров.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ActorWithMovies"
                            }
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "description": "JSON-структура ответа с сообщением об ошибке.",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.ActorWithMovies": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "Birth date of the actor",
                    "type": "string"
                },
                "gender": {
                    "description": "Gender of the actor",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier for the actor",
                    "type": "integer"
                },
                "movies": {
                    "description": "Movies associated with the actor",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Movie"
                    }
                },
                "name": {
                    "description": "Name of the actor",
                    "type": "string"
                }
            }
        },
        "model.InputActor": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "Birth date of the actor",
                    "type": "string"
                },
                "gender": {
                    "description": "Gender of the actor",
                    "type": "string"
                },
                "movies": {
                    "description": "Movies associated with the actor",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Movie"
                    }
                },
                "name": {
                    "description": "Name of the actor",
                    "type": "string"
                }
            }
        },
...
        "model.Movie": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of the movie",
                    "type": "string"
                },
                "rating": {
                    "description": "Rating of the movie",
                    "type": "integer"
                },
                "release_date": {
                    "description": "Release date of the movie",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the movie",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}